TEST: Move test_future
