+++ psychopy/tests/test_experiment/needs_wx/test_Experiment.py
@@ -79,11 +79,6 @@ class TestExpt():
         shutil.rmtree(cls.tmp_dir, ignore_errors=True)
 
 
-        schema_name = path.join(self.exp.prefsPaths['psychopy'], 'experiment', 'experiment.xsd')
-        schema = xmlschema.XMLSchema(schema_name)
-
-        for psyexp_file in psyexp_files:
-            assert schema.is_valid(psyexp_file), f"Error in {psyexp_file}:\n" + "\n".join(err.reason for err in schema.iter_errors(psyexp_file))
 
     def _checkLoadSave(self, file):
         exp = self.exp
@@ -106,15 +101,7 @@ class TestExpt():
 
         return py_file, psyexp_file
 
-    def _checkCompile(self, py_file):
-        # compile the temp file to .pyc, catching error msgs
-        # (including no file at all):
-        try:
-            py_compile.compile(py_file, doraise=True)
-        except py_compile.PyCompileError as err:
-            err.msg = py_file
-            raise err
-        return py_file + 'c'
+
 
     def _checkPyDiff(self, file_py, file2_py):
         """return '' for no meaningful diff, or a diff patch"""
@@ -228,22 +215,6 @@ class TestExpt():
         assert not diff_in_file_py ### see known_py_diffs.txt; potentially a locale issue? ###
         #assert not diff_in_file_psyexp # was failing most times, uninformative
         #assert not diff_in_file_pyc    # oops, was failing every time despite identical .py file
-
-    def test_future(self):
-        """An experiment file with made-up params and routines to see whether
-        future versions of experiments will get loaded.
-        """
-        expfile = path.join(self.exp.prefsPaths['tests'], 'data', 'futureParams.psyexp')
-        self.exp.loadFromXML(expfile) # reload the edited file
-        # we don't test this script but make sure it builds
-        script = self.exp.writeScript(expPath=expfile)
-        py_file = os.path.join(self.tmp_dir, 'testFutureFile.py')
-        # save the script:
-        with codecs.open(py_file, 'w', 'utf-8-sig') as f:
-            f.write(script)
-
-        #check that files compiles too
-        self._checkCompile(py_file)
 
     def test_loopBlocks(self):
         """An experiment file with made-up params and routines to see whether
